feat: improve testing of the installation of dependencies (#10)
* feat: add tests\dependencies.Tests.ps1

Add tests to check that dependencies are installed. It is different from
the tests\Install-Dependencies.Tests.ps1 that tests that
src\Install-Depenencies.ps1 is working correctly

* feat: refactor Install-Dependencies.Tests.ps1

Refactor the tests\Install-Dependencies.Tests.ps1 to use the
tests\dependencies.Tests.ps1 to check that dependencies are installed.

* chore: rename test

Rename the name of the tests to something more comprehensible